# perf(memory_usage): reduce the latency of memory usage fetching in linux.