--- a/Cargo.toml
+++ b/Cargo.toml
@@ -26,14 +26,13 @@ http = ["attohttpc"]
 # Vendor OpenSSL, use this if you have trouble cross-compiling starship
 tls-vendored = ["native-tls/vendored"]
 
-
 [dependencies]
 clap = "2.33.3"
 ansi_term = "0.12.1"
 dirs-next = "2.0.0"
 git2 = { version = "0.13.12", default-features = false }
 toml = { version = "0.5.7", features = ["preserve_order"] }
-rust-ini = "0.16" 
+rust-ini = "0.16"
 serde_json = "1.0.59"
 rayon = "1.5.0"
 log = { version = "0.4.11", features = ["std"] }
@@ -45,8 +44,6 @@ unicode-segmentation = "1.6.0"
 gethostname = "0.2.1"
 once_cell = "1.5.1"
 chrono = "0.4"
-sysinfo = "0.15.3"
-procfs = "0.8.1"
 byte-unit = "4.0.9"
 starship_module_config_derive = { version = "0.1.2", path = "starship_module_config_derive" }
 yaml-rust = "0.4"
@@ -63,11 +60,16 @@ rand = "0.7.3"
 serde_derive = "1.0.115"
 indexmap = "1.6.0"
 notify-rust = { version = "4.0.0", optional = true }
-
 # Optional/http:
-attohttpc = { version = "0.16.0", optional = true, default-features = false, features = ["tls", "form"] }
+attohttpc = { version = "0.16.0", optional = true, default-features = false, features = [
+  "tls",
+  "form"
+] }
 native-tls = { version = "0.2", optional = true }
 shell-words = "1.0.0"
+
+[target.'cfg(target_os = "linux")'.dependencies]
+procfs = "0.8.1"
 
 [target.'cfg(windows)'.dependencies]
 winapi = { version = "0.3", features = [
@@ -77,6 +79,9 @@ winapi = { version = "0.3", features = [
   "handleapi",
   "impl-default",
 ] }
+
+[target.'cfg(not(target_os = "linux"))'.dependencies]
+sysinfo = "0.15.3"
 
 [target.'cfg(not(windows))'.dependencies]
 nix = "0.19.0"