--- a/src/modules/utils/meminfo/meminfo_general.rs
+++ b/src/modules/utils/meminfo/meminfo_general.rs
@@ -1,5 +1,4 @@
 use super::MemoryUsage;
-use std::ops::Deref;
 use sysinfo::{RefreshKind, System, SystemExt};
 pub struct MemInfo(System);