--- a/src/modules/utils/mod.rs
+++ b/src/modules/utils/mod.rs
@@ -1,4 +1,5 @@
 pub mod directory;
+pub mod meminfo;
 
 #[cfg(target_os = "windows")]
 pub mod directory_win;