<!--- Provide a general summary of your changes in the Title above -->
<!--- To help with semantic versioning the PR title should start with one of the conventional commit types. -->
<!--- The conventional commit types for Semantic PR are: feat, fix, docs, style, refactor, perf, test, build, ci, chore, revert -->

#### Description
<!--- Describe your changes in detail -->
This pull request uses [`procfs`](https://docs.rs/procfs/0.8.1) to fetch memory usage in Linux, instead of [`sysinfo`](https://docs.rs/sysinfo/0.15.3) which will refresh some other fields while we just try to refresh memory info.

Here is a [benchmark](https://gist.github.com/iosmanthus/75d192ba0cb2ed45b2f3c04a4c24973e) of these two crates comparing with meminfo fetching:

```rust
    Finished bench [optimized] target(s) in 0.10s
     Running target/release/deps/meminfo-33685aca30a53811

running 2 tests
test benches::bench_procfs  ... bench:      31,920 ns/iter (+/- 7,817)
test benches::bench_sysinfo ... bench:     261,627 ns/iter (+/- 51,030)

test result: ok. 0 passed; 0 failed; 0 ignored; 2 measured; 0 filtered out

```

#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->
In my environment, the fish prompt runs really slow while using `sysinfo`:
```sh
❯ starship timings
 Here are the timings of modules in your prompt (>=1ms or output):
 memory_usage  -  435ms  -   using 🐏 56GiB/126GiB = 45% <--- about 500ms
 username      -    4ms  -   🎃 iosmanthus
 hostname      -   <1ms  -   ip-172-16-5-32 in
 directory     -   <1ms  -   ~
 line_break    -   <1ms  -   \n
 time          -   <1ms  -   🕙 19:59:42
 character     -   <1ms  -   ❯
```

While using `procfs` instead, the latency drops tremendously:
```sh
 Here are the timings of modules in your prompt (>=1ms or output):
 username      -   5ms  -   🎃 iosmanthus
 hostname      -  <1ms  -   ip-172-16-5-32 in
 directory     -  <1ms  -   ~
 memory_usage  -  <1ms  -   using 🐏 56GiB/126GiB = 45% <--- less than 1ms
 line_break    -  <1ms  -   \n
 time          -  <1ms  -   🕙 20:18:56
 character     -  <1ms  -   ❯
```

This pull request also introduces a `meminfo` module in `modules/utils` which use a trait `MemoryUsgae` and a struct `MemInfo` to abstract the behaviour of different memory information fetching approaches which should work with other pull requests like https://github.com/starship/starship/pull/1573

#### Screenshots (if appropriate):
Works fine in Windows, passed compilation:
![image](https://user-images.githubusercontent.com/16307070/99151077-a437b000-26d3-11eb-953f-b3cf87420e32.png)


#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, tests ran to see how -->
<!--- your change affects other areas of the code, etc. -->
- [ ] I have tested using **MacOS**
- [x] I have tested using **Linux**
- [x] I have tested using **Windows**

#### Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [ ] I have updated the documentation accordingly.
- [ ] I have updated the tests accordingly.