@davidkna 2020-11-15
I think it might make sense to just replace `sysinfo` with another library. For example [`sys-info`](https://crates.io/crates/sys-info) looks like a decent alternative with good platform support.
Beyond that I'm not sure if adding a library for procfs is necessary, parsing procfs seems to amount to `split_whitespace` and maybe handling the units.